factor built-in attribute parsing into submodule
